10.1.5
deleted random.test.js and made a new file called Potion.test.js

const Potion = require("../lib/Potion.js");
This means it is looking for any exported functions from the Potion.js file

test("creates a health potion object", () => {
  const potion = new Potion("health");
This we use test to set up our code and use new to make sure this variable creates a new potion object.  This will be explained more later.

  expect(potion.name).toBe("health");
  expect(potion.value).toEqual(expect.any(Number));
});
These expects  are looking at the new potion made in the test function and seeing if it's name and it's value match.

Afterwards we ran a test expecting it to fail because we have yet to add any code to Potion.js.