10.1.6
Did some testing to see how constructors work using cars.  Worth reading I get lost again.

Then we built this in potion.js.

function Potion(name) {
This takes in the value of the variable potion and sets it to name.

  this.name = name;
Then this creates a potion.name using this.name to set the value of the name of the potion to it's name in the object.

  if (this.name === "health") {
this checks if the value we passed in is called "health"  if it is

    this.value = Math.floor(Math.random() * 10 + 30);
it creates a value variable in the object  and sets it's number to somewhere between 30-40.

  } else {
    this.value = Math.floor(Math.random() * 5 + 7);
otherwise it sets the value of this potion to a value between 5-12.
  }
}

module.exports = Potion;
Finally we export this Potion function out where Potion.test.js will be looking for it.