10.1.7
created a test function that check for a random potion in Potiontest.js

test("creates a random potion object", () => {
  const potion = new Potion();
creates a potion with no value being passed in

  expect(potion.name).toEqual(expect.any(String));
then checks to make sure what gets sent back is a string
  expect(potion.name.length).toBeGreaterThan(0);
makes sure the string is not empty
  expect(potion.value).toEqual(expect.any(Number));
and that the value attribute as any number
});

Then in Potion.js we made this

function Potion(name) {
  this.types = ["strength", "agility", "health"];
created an array to handle this next bit
  this.name = name || this.types[Math.floor(Math.random() * this.types.length)];
this checks to see if name has something passed in it.  If it does it uses that.  Otherwise it looks at this.types array that we just made and grabs a random name and assigns it.

  if (this.name === "health") {
    this.value = Math.floor(Math.random() * 10 + 30);
  } else {
    this.value = Math.floor(Math.random() * 5 + 7);
  }
}

module.exports = Potion;

The rest functions as before.