10.2.4
added a function to Player.js to make this work

function Player(name = "") {
sets name to an empty string if no name is provided with ES6 magic.

  this.name = name;
sets the name to the passed in name Dave.

  this.health = Math.floor(Math.random() * 10 + 95);
health can be between 95-105

  this.strength = Math.floor(Math.random() * 5 + 7);
  this.agility = Math.floor(Math.random() * 5 + 7);
these can be between 7-12
}

module.exports = Player;
kicks it out where Player.test.js is waiting for it.