10.2.7
added

Player.prototype.getStats = function () {
  return {
    potions: this.inventory.length,
    health: this.health,
    strength: this.strength,
    agility: this.agility,
  };
};

Player.prototype.getInventory = function () {
  if (this.inventory.length) {
    return this.inventory;
  }
  return false;
};

to Player.js  to make it so this test passes.  My best guess is that they are now being called and returning properties with nothing in it.  But since the properties exist then the test in Player.test.js is ok  with this.  They kind of glossed over this.  Hopefully there will be more explained in the future.