10.4.4
Added the startNewBattle prototype method.

Game.prototype.startNewBattle = function () {
  if (this.player.agility > this.currentEnemy.agility) {
    this.isPlayerTurn = true;
  } else {
    this.isPlayerTurn = false;
  }
This is checking to see who's turn it is which is not relevant yet.
  console.log("Your stats are as follows:");
  console.table(this.player.getStats());
  console.log(this.currentEnemy.getDescription());
These console logs are pretty neat, I met console.table for the first time which would have been good for yesterday.

  this.battle();
This is a function that breaks things because we have not added it yet.
};