10.4.5
used inquirer to get data from a list in the battle method.

Game.prototype.battle = function () {
  if (this.isPlayerTurn) {
    inquirer
      .prompt({
        type: "list",
        message: "What would you like to do?",
        name: "action",
        choices: ["Attack", "Use potion"],
      })
Uses Inquirer to make a list to ask if you want to use a potion or attack.

      .then(({ action }) => {
        if (action === "Use potion") {
          if (!this.player.getInventory()) {
            console.log("You don't have any potions!");
            return;
          }
Then it references it's name "action" and asks if we selected Use Potion.  If we do it first checks if our getInventory is returning a false.  If so It console.logs You don't have any potions and kicks us out.

          inquirer
            .prompt({
              type: "list",
              message: "Which potion would you like to use?",
              name: "action",
              choices: this.player
                .getInventory()
                .map((item, index) => `${index + 1}: ${item.name}`),
            })
Otherwise inquirer will ask us another list question using fancy .map action to get our actual inventory and then display it like this 1: Health Potion.
            .then(({ action }) => {
              const potionDetails = action.split(": ");

              this.player.usePotion(potionDetails[0] - 1);
              console.log(`You used a ${potionDetails[1]} potion.`);
            });
Finally we is this .then to call the usePotion we use split to get the number vale from 1: Health to make it look like this [1, Health] and use the usePotion method and target that number giving it a -1 to counter our previous +1 we created for the user to read.   Then we console log the name of the potion in that string by targeting the second spot in that array.
        } else {
          const damage = this.player.getAttackValue();
          this.currentEnemy.reduceHealth(damage);

          console.log(`You attacked the ${this.currentEnemy.name}`);
          console.log(this.currentEnemy.getHealth());
        }
      });
  } else {
    const damage = this.currentEnemy.getAttackValue();
    this.player.reduceHealth(damage);

    console.log(`You were attacked by the ${this.currentEnemy.name}`);
    console.log(this.player.getHealth());
otherwise we just smack the enemy in a pretty straight forward way.
  }
};