10.5.6
We met the extends keyword which allows a class to talk to another class.  For example....
class Enemy extends Character {
  constructor(name, weapon) {
Will have the Enemy class get data from the Character Class as long as....
super(name);
The word super() call is added in to talk to the parent Class.  We also moved moved duplicate properties out of the Player and Enemy Classes and put them in the Character Constructor and then was able to pass them in into using super(name)